add more mongo AST rendering tests
